+++ ldap2pg/manager.py
@@ -235,7 +235,7 @@ class SyncManager(object):
             grant = mapping.get('grant', [])
             aclitems = self.apply_grant_rules(grant, dbname, schema, entries)
             for aclitem in aclitems:
-                if aclitem.dbname not in ['__all__'] + databases:
+                if aclitem.dbname not in [AclItem.ALL_DATABASES] + databases:
                     msg = "Database %s does not exists or is not managed." % (
                         aclitem.dbname,)
                     raise UserError(msg)