+++ docs/changelog.md
@@ -19,6 +19,8 @@ follow
 - Show detailed version informations.
 - Show YAML parsing error.
 - Fix various configuration loading errors.
+- Quote role name in SQL queries.
+- Documentation and sample update.
 
 
 # ldap2pg 3.3